fixed tests for x86_32; fixed warnings; generate exe for 32bits on x86_32 and for 64bits on x86_64

+++ tests/mock/mock_paa.c
@@ -14,7 +14,7 @@ uint32 PAA_alloc(uint64 a_size, uint64 *a_address, uint64 a_alignment)
     }
     *a_address = 0;
 
-    uint64 g_placementAddress = (uint64) malloc(a_size + a_alignment - 1);
+    uint64 g_placementAddress = (uint64) (size_t) malloc(a_size + a_alignment - 1);
     if ((g_placementAddress & (a_alignment - 1)) != 0) {
         g_placementAddress &= (~(a_alignment - 1));
         g_placementAddress += a_alignment;
